mind: validate authored event batches (#15)
Add pure validation for model proposals with canonical subject ordering, unknown-human filtering, deterministic batch truncation, signed permille clamping, and named per-field effect bounds. Valid proposals map only onto the safe event vocabulary already owned by core.\n\nVerified: cargo fmt --all; cargo clippy --all-targets --all-features -- -D warnings; cargo test --workspace.

+++ crates/mind/src/types.rs
@@ -133,9 +133,9 @@ pub struct AiEventProposal {
     pub kind: AiEventKind,
     pub title: String,
     pub description: String,
-    pub base_prob: Option<u16>,
+    pub base_prob: Option<i64>,
     pub skill_modifier: Option<SkillId>,
-    pub modifier_strength: Option<u16>,
+    pub modifier_strength: Option<i64>,
     pub effects: Vec<AiEffect>,
     pub seed_salt: u32,
 }
